requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
awscli==1.17.2
beautifulsoup4==4.8.2
botocore==1.14.2
cachetools==3.0.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
distlib==0.3.0
Django==2.1.3
django-environ==0.4.5
django-mysql==3.3.0
djangorestframework==3.11.0
docutils==0.15.2
filelock==3.0.12
google-api-python-client==1.7.4
google-auth==1.6.1
google-auth-httplib2==0.0.3
httplib2==0.12.0
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
multidict==4.7.5
mysqlclient==1.4.6
oauth2client==4.1.3
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pylint==2.4.4
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.8.1
pytz==2018.7
PyYAML==5.2
requests==2.23.0
rsa==4.0
s3transfer==0.3.0
six==1.11.0
slackbot==0.5.6
slackclient==2.5.0
slacker==0.14.0
soupsieve==2.0
typed-ast==1.4.1
uritemplate==3.0.0
urllib3==1.25.7
uWSGI==2.0.17.1
virtualenv==20.0.15
wcwidth==0.1.9
websocket-client==0.44.0
wrapt==1.11.2
yarl==1.4.2
zipp==3.1.0
